Change 'CONTACT' to 'Contact Theatre'
I think the theatre use the capitalised word "CONTACT" as their
logo, but often go by the more descriptive "Contact Theatre". For
example, they use "Contact Theatre" in the screen reader-aware h1
tag on their home page, and in the title of their YouTube channel.

"Contact Theatre" is also how the venue is identified on Google Maps, so
I imagine this'll help people who are planning their trip.

I hope I'm not coming across as a pedant here, but (as a Manchester
local) I didn't know what CONTACT referred to and didn't try searching
for "contact Manchester". As soon as I knew it was a theatre I knew how
to look it up.