Drop suggestion people bring their own bike
I'm not sure about the secure bike parking (maybe it's great) but
without seeing it I'm a bit nervous about this. I never lock my bike up
in the city for more than a few minutes.

If Contact's bike storage involves a locked door we should mention it,
but I won't be able to look into that in time before this is put live.