Remove schedule layout hack
This was put in because (IIRC) the first room lined up with the edge of the screen
at Cardiff, and made it appear there was only one track when viewing on mobile.
Nowadays, Pretalx automatically adjusts itself to show everything in a single column,
and this hack therefore doesn't do anything helpful, but does mean that the date
header slowly disappears from view in a parallax sort of way as one scrolls down.